refactor(workflows): extract WorkflowDetailsDrawer from canvas-editor (phase 2)
Pulls the self-contained workflow name/description drawer out of
workflow-canvas-editor.tsx into its own one-component file, following the
sibling pattern (node-config-drawer.tsx, step-detail-drawer.tsx).

WorkflowDetailsDrawer owns its draft name/description state (re-seeded from
the definition prop), has a props-only interface, and closes over no editor
state — a clean lift. The component imports it back and sheds the now-unused
Input / Label / Textarea / X imports (the drawer was their only consumer).
1,469 lines remaining.

Pure behavior-preserving relocation — the JSX is byte-identical.

Verified: typecheck + eslint + workflow-canvas-editor.test (29/0) + full suite
(5124/0) + binary build + isolated-home boot smoke (Workflows loads; rebuilt
client.js → 200).
