+++ plugins/workflows/components/workflow-canvas-editor.tsx
@@ -55,12 +55,8 @@ import {
   Trash2,
   UserRound,
   Workflow as WorkflowIcon,
-  X,
 } from 'lucide-react'
 import { Button } from "@makinbakin/sdk/ui"
-import { Input } from "@makinbakin/sdk/ui"
-import { Label } from "@makinbakin/sdk/ui"
-import { Textarea } from "@makinbakin/sdk/ui"
 import {
   Dialog,
   DialogContent,
@@ -73,6 +69,7 @@ import {
 import { getNodeRendererSnapshot, subscribeNodeRenderers } from '../lib/node-renderer-registry'
 import { NodeTypePalette, PALETTE_DRAG_MIME_TYPE } from './node-type-palette'
 import { NodeConfigDrawer } from './node-config-drawer'
+import { WorkflowDetailsDrawer } from './workflow-details-drawer'
 import { ManagedWorkflowCopyDialog } from './managed-workflow-copy-dialog'
 import { AgentAssignmentLabel } from './nodes/agent-assignment-label'
 import {
@@ -294,98 +291,6 @@ const BUILTIN_NODE_TYPES: NodeTypes = {
   trigger: EditorTriggerNode,
   subflowGroup: EditorSubflowGroupNode,
   appendStep: EditorAppendStepNode,
-}
-
-function WorkflowDetailsDrawer({
-  definition,
-  onApply,
-  onClose,
-  applyLabel = 'Apply',
-  applying = false,
-}: {
-  definition: WorkflowDefinition
-  onApply: (patch: Pick<WorkflowDefinition, 'name' | 'description'>) => void | Promise<void>
-  onClose: () => void
-  applyLabel?: string
-  applying?: boolean
-}) {
-  const [name, setName] = useState(definition.name)
-  const [description, setDescription] = useState(definition.description ?? '')
-  const canApply = name.trim().length > 0
-
-  useEffect(() => {
-    setName(definition.name)
-    setDescription(definition.description ?? '')
-  }, [definition.description, definition.name])
-
-  return (
-    <aside className="flex w-[25rem] flex-col border-l border-border bg-card">
-      <div className="flex items-start justify-between gap-3 border-b border-border p-3">
-        <div className="min-w-0">
-          <h2 className="truncate text-sm font-medium">Workflow details</h2>
-          <p className="mt-0.5 text-xs text-muted-foreground">
-            Edit the workflow name and description.
-          </p>
-        </div>
-        <Button
-          type="button"
-          variant="ghost"
-          size="icon-sm"
-          aria-label="Close workflow details"
-          onClick={onClose}
-        >
-          <X className="size-3.5" />
-        </Button>
-      </div>
-      <div className="flex-1 overflow-y-auto p-3">
-        <div className="mb-3">
-          <Label className="text-xs" htmlFor="workflow-details-name">
-            Name <span className="text-red-400">*</span>
-          </Label>
-          <Input
-            id="workflow-details-name"
-            value={name}
-            onChange={(event) => setName(event.target.value)}
-            placeholder="Workflow name"
-            aria-invalid={!canApply || undefined}
-          />
-          {!canApply && (
-            <p className="mt-1 text-[11px] font-medium text-red-300">Enter a workflow name.</p>
-          )}
-        </div>
-        <div className="mb-3">
-          <Label className="text-xs" htmlFor="workflow-details-description">
-            Description
-          </Label>
-          <Textarea
-            id="workflow-details-description"
-            rows={5}
-            value={description}
-            onChange={(event) => setDescription(event.target.value)}
-            placeholder="Describe when this workflow should be used"
-          />
-        </div>
-      </div>
-      <div className="flex items-center justify-between gap-2 border-t border-border p-3">
-        <Button size="sm" variant="ghost" onClick={onClose} disabled={applying}>
-          Cancel
-        </Button>
-        <Button
-          size="sm"
-          onClick={async () => {
-            if (!canApply) return
-            await onApply({
-              name: name.trim(),
-              description: description.trim(),
-            })
-          }}
-          disabled={!canApply || applying}
-        >
-          {applying ? 'Saving...' : applyLabel}
-        </Button>
-      </div>
-    </aside>
-  )
 }
 
 interface WorkflowCanvasEditorProps {
