+++ plugins/workflows/components/workflow-details-drawer.tsx
@@ -1,0 +1,106 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import { X } from 'lucide-react'
+import { Button } from "@makinbakin/sdk/ui"
+import { Input } from "@makinbakin/sdk/ui"
+import { Label } from "@makinbakin/sdk/ui"
+import { Textarea } from "@makinbakin/sdk/ui"
+import type { WorkflowDefinition } from '../types'
+
+/**
+ * Controlled drawer for editing a workflow's name + description. Self-contained:
+ * owns its draft name/description state (re-seeded when the definition prop
+ * changes), props-only interface, no closure over editor state.
+ */
+export function WorkflowDetailsDrawer({
+  definition,
+  onApply,
+  onClose,
+  applyLabel = 'Apply',
+  applying = false,
+}: {
+  definition: WorkflowDefinition
+  onApply: (patch: Pick<WorkflowDefinition, 'name' | 'description'>) => void | Promise<void>
+  onClose: () => void
+  applyLabel?: string
+  applying?: boolean
+}) {
+  const [name, setName] = useState(definition.name)
+  const [description, setDescription] = useState(definition.description ?? '')
+  const canApply = name.trim().length > 0
+
+  useEffect(() => {
+    setName(definition.name)
+    setDescription(definition.description ?? '')
+  }, [definition.description, definition.name])
+
+  return (
+    <aside className="flex w-[25rem] flex-col border-l border-border bg-card">
+      <div className="flex items-start justify-between gap-3 border-b border-border p-3">
+        <div className="min-w-0">
+          <h2 className="truncate text-sm font-medium">Workflow details</h2>
+          <p className="mt-0.5 text-xs text-muted-foreground">
+            Edit the workflow name and description.
+          </p>
+        </div>
+        <Button
+          type="button"
+          variant="ghost"
+          size="icon-sm"
+          aria-label="Close workflow details"
+          onClick={onClose}
+        >
+          <X className="size-3.5" />
+        </Button>
+      </div>
+      <div className="flex-1 overflow-y-auto p-3">
+        <div className="mb-3">
+          <Label className="text-xs" htmlFor="workflow-details-name">
+            Name <span className="text-red-400">*</span>
+          </Label>
+          <Input
+            id="workflow-details-name"
+            value={name}
+            onChange={(event) => setName(event.target.value)}
+            placeholder="Workflow name"
+            aria-invalid={!canApply || undefined}
+          />
+          {!canApply && (
+            <p className="mt-1 text-[11px] font-medium text-red-300">Enter a workflow name.</p>
+          )}
+        </div>
+        <div className="mb-3">
+          <Label className="text-xs" htmlFor="workflow-details-description">
+            Description
+          </Label>
+          <Textarea
+            id="workflow-details-description"
+            rows={5}
+            value={description}
+            onChange={(event) => setDescription(event.target.value)}
+            placeholder="Describe when this workflow should be used"
+          />
+        </div>
+      </div>
+      <div className="flex items-center justify-between gap-2 border-t border-border p-3">
+        <Button size="sm" variant="ghost" onClick={onClose} disabled={applying}>
+          Cancel
+        </Button>
+        <Button
+          size="sm"
+          onClick={async () => {
+            if (!canApply) return
+            await onApply({
+              name: name.trim(),
+              description: description.trim(),
+            })
+          }}
+          disabled={!canApply || applying}
+        >
+          {applying ? 'Saving...' : applyLabel}
+        </Button>
+      </div>
+    </aside>
+  )
+}
